+++ internal/forge/gh/proposal_finder.go
@@ -10,6 +10,18 @@ import (
 	. "github.com/git-town/git-town/v21/pkg/prelude"
 )
 
+func (self Connector) FindProposalFn() Option[func(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)] {
+	return Some(self.findProposal)
+}
+
+func (self Connector) findProposal(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	out, err := self.Backend.Query("gh", "pr", "list", "--head="+branch.String(), "--base="+target.String(), "--json=number,title,body,mergeable,headRefName,baseRefName,url")
+	if err != nil {
+		return None[forgedomain.Proposal](), err
+	}
+	return ParseJSONOutput(out, branch)
+}
+
 func ParseJSONOutput(output string, branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
 	var parsed []jsonData
 	err := json.Unmarshal([]byte(output), &parsed)
